New DEM solver and programmed a drop of a ball

+++ Version 0.0.4/Scene.h
@@ -51,7 +51,7 @@ public:
 	double normalStiffness = 1e6;
 	double shearStiffness  = 0.5e6;
 
-	double simDuration = 10000.0;
+	int simDuration = 10000;
 	double Time = 0.0;
 	int subCycleNumber = 0;
